makes post queries handled by a helper function
