+++ react/src/pages/CodingPracticePage.js
@@ -1,17 +1,15 @@
 import React, { useContext } from "react";
 import { Page } from "../components/Page";
 import styles from "../styles/main.scss";
-import { useLocation } from "react-router-dom";
 import MasterContext from "../reducer/context";
 
 const { codingPracticeHeader, codingPracticeTitle, codingPracticeContents } = styles;
 const CodingPracticePage = (props) => {
     const { dispatch } = useContext(MasterContext);
-    const location = useLocation();
     return (
         <Page {...props}>
             <div className={codingPracticeHeader}>
-                CodingPractice (location: {location.pathname})
+                CodingPractice
             </div>
             <div className={codingPracticeContents}>
                 <div className={codingPracticeTitle}>
