+++ react/src/util/helpers.js
@@ -1,3 +1,7 @@
+import serverURL from "./serverinfo";
+import axios from "axios";
+import https from "https";
+
 /**
  * 
  * @param {Object} props The props Object passed to functional React components
@@ -12,8 +16,28 @@ export const parseProps = (props, validProps) => {
         }
     }
     return style;
-}
+};
 
 export const objIsEmpty = (obj) => {
     return obj && Object.keys(obj).length === 0;
-}
+};
+
+export const queryServer = (useCase, data, successCallback, errorCallback) => {
+    const postData = {
+        use: useCase,
+        ...data,
+    };
+
+    axios.post(serverURL, postData, {
+		headers: {
+			'Content-Type': 'application/json',
+		},
+		timeout: 1000,
+		withCredentials: true,
+		httpsAgent: new https.Agent({ keepAlive: true }),
+	}).then((res) => {
+        successCallback(res);
+    }).catch((error) => {
+        errorCallback(error);
+    });
+};
